More indirection removal.
Centralized all the global vars (which are mostly
"constants" to the extent that go-lang makes that possible,
which is to say, none) in suffuse.go.
