+++ cmd/suffuse/main.go
@@ -2,13 +2,10 @@ package main
 
 import . "github.com/suffuse/go-suffuse/suffuse"
 
-var mfs Sfs
-var err error
-
 func main() {
   opts := ParseSfsOpts()
 
-  mfs, err = opts.Create()
+  mfs, err := opts.Create()
   MaybePanic(err)
 
   err = mfs.Serve()
